move simulation to its own module
